recovery

+++ ft_printf/ft_printf.h
@@ -3,15 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: arigonza <arigonza@student.42malaga.com>   +#+  +:+       +#+        */
+/*   By: arigonza < arigonza@student.42malaga.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 13:35:27 by arigonza          #+#    #+#             */
-/*   Updated: 2023/01/27 13:49:52 by arigonza         ###   ########.fr       */
+/*   Updated: 2023/02/19 11:41:36 by arigonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
-#define FT_PRINTF_H
+# define FT_PRINTF_H
 
 # include <stdarg.h>
 # include <unistd.h>
@@ -19,12 +19,12 @@
 # include <stdlib.h>
 # include <stdio.h>
 
-int	ft_printf(const char *string, ...);
+int		ft_printf(const char *string, ...);
 void	ft_putchar(int c, int *count);
 void	ft_putstr(char *str, int *count);
 void	ft_putnbr(int n, int *count);
 void	ft_putunsignednbr(unsigned int nbr, int *count);
-void	ft_printhexa(unsigned long nbr, char h, int *count);
+void	ft_printhexa(unsigned long long nbr, char h, int *count);
 void	ft_printptr(void *ptr, int *count);
 
 #endif
